requirements.txt
flask
gunicorn
sqlalchemy
pyodbc
python-dotenv
PyPDF2
nltk
upstash-redis
groq
flask-jwt-extended
supabase
pinecone
numpy>=1.26.0
pymysql
email-validator
pydantic
gradio_client
flask-cors
